encapsulate task dir and move test related code to test file
