+++ pkg/planner/memo/BUILD.bazel
@@ -8,8 +8,6 @@ go_library(
         "group_expr.go",
         "implementation.go",
         "pattern.go",
-        "task.go",
-        "task_scheduler.go",
     ],
     importpath = "github.com/pingcap/tidb/pkg/planner/memo",
     visibility = ["//visibility:public"],
@@ -29,12 +27,10 @@ go_test(
         "group_test.go",
         "main_test.go",
         "pattern_test.go",
-        "task_scheduler_test.go",
-        "task_test.go",
     ],
     embed = [":memo"],
     flaky = True,
-    shard_count = 25,
+    shard_count = 22,
     deps = [
         "//pkg/domain",
         "//pkg/expression",
