+++ pkg/planner/memo/task/task_scheduler.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package memo
+package task
 
 var _ TaskScheduler = &SimpleTaskScheduler{}
